try to disable testsuite for ghc7
Resolving dependencies...
cabal: Cannot test the package lsfrom-0.1 because none of the components are
available to build: the test suite 'test' is not available because the solver
did not find a plan that included the test suites. Force the solver to enable
this for all packages by adding the line 'tests: True' to the
'cabal.project.local' file.